pyproject.toml
[project]
name = "fast-doctolib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard,standart]>=0.116.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.25",
]
